Improve the ``PortBindingUpdateVirtualPortsEvent`` match filter
This patch improves the ``PortBindingUpdateVirtualPortsEvent``
match filter. These are the new conditions:
* Event delete: that happens when the port binding has been deleted
  because the port is no longer bound or the port has been
  deleted. That will remove the Neutron port host name.
  NOTE: in case the Neutron port has been deleted, the method
  ``Ml2Plugin.update_virtual_port_host`` won't update (create) a new
  PortBinding object.
* If the new register has virtual_parents but not the old one, that
  means the ovn-controller has received traffic with the VIP from
  this port. The port host ID must be set.
* If the virtual parents have changed, the port host ID must be
  updated.
* If the virtual parents have been removed, the port host ID must
  be removed too.

Newer versions of OVN [1] are handling the virtual port binding in
a different way. When the virtual parents are added or removed,
the related "Port_Binding" register is deleted and the created
again. This is why this new version includes the event "DELETE"
on the match method; when the register is deleted, the event
class considers that the port is no longer bound to a host and
removes the host name for the Neutron port.

[1]https://review.opendev.org/c/openstack/neutron/+/880890/

+++ neutron/plugins/ml2/plugin.py
@@ -2068,8 +2068,14 @@ class Ml2Plugin(db_base_plugin_v2.NeutronDbPluginV2,
         """
         hostname = hostname or ''
         with db_api.CONTEXT_WRITER.using(context):
-            for pb in ports_obj.PortBinding.get_objects(context,
-                                                        port_id=port_id):
+            pbindings = ports_obj.PortBinding.get_objects(context,
+                                                          port_id=port_id)
+            if not pbindings:
+                # The port has been deleted and there is no need to delete and
+                # create any port binding.
+                return
+
+            for pb in pbindings:
                 pb.delete()
 
             attrs = {'port_id': port_id,
